bkfe {KernSmooth}	R Documentation

Compute a Binned Kernel Functional Estimate

Description

Returns an estimate of a binned approximation to the kernel estimate of the specified density functional. The kernel is the standard normal density.

Usage

bkfe(x, drv, bandwidth, gridsize = 401L, range.x, binned = FALSE,
     truncate = TRUE)

Arguments

x 	 vector of observations from the distribution whose density is to be estimated. Missing values are not allowed.

drv 	 order of derivative in the density functional. Must be a non-negative even integer.

bandwidth 	 the kernel bandwidth smoothing parameter. Must be supplied.

gridsize 	 the number of equally-spaced points over which binning is performed.

range.x 	 vector containing the minimum and maximum values of x at which to compute the estimate. The default is the minimum and maximum data values, extended by the support of the kernel.

binned 	 logical flag: if TRUE, then x and y are taken to be grid counts rather than raw data.

truncate 	 logical flag: if TRUE, data with x values outside the range specified by range.x are ignored.

Details

The density functional of order drv is the integral of the product of the density and its drvth derivative. The kernel estimates of such quantities are computed using a binned implementation, and the kernel is the standard normal density.

Value

the (scalar) estimated functional.

Background

Estimates of this type were proposed by Sheather and Jones (1991).

References

Sheather, S. J. and Jones, M. C. (1991). A reliable data-based bandwidth selection method for kernel density estimation. Journal of the Royal Statistical Society, Series B, 53, 683–690.

Wand, M. P. and Jones, M. C. (1995). Kernel Smoothing. Chapman and Hall, London.

Examples

data(geyser, package="MASS")
x <- geyser$duration
est <- bkfe(x, drv=4, bandwidth=0.3)

[Package KernSmooth version 2.23-2 Index]